Add test for TaskUtils task executor service
A new test method has been added to check the functionality of the TaskUtils task executor service. This test evaluates if no exception is thrown when the task is completed within the expected time duration and also checks if a TimeoutException is thrown when the task exceeds the specified timeout.